refactor Asset/Token named things, separate the claritydb struct from the k/v wrapper code
