+++ src/vm/checker/errors.rs
@@ -27,12 +27,11 @@ pub enum CheckErrors {
     CheckerImplementationFailure,
 
     // Assets
-    BadAssetName,
     BadTokenName,
-    DefineTokenBadSignature,
-    DefineAssetBadSignature,
-    NoSuchAsset(String),
-    NoSuchToken(String),
+    DefineFTBadSignature,
+    DefineNFTBadSignature,
+    NoSuchNFT(String),
+    NoSuchFT(String),
 
     // tuples
     BadTupleFieldName,
@@ -211,12 +210,11 @@ impl DiagnosableError for CheckErrors {
             CheckErrors::IllegalOrUnknownFunctionApplication(function_name) => format!("use of illegal / unresolved function '{}", function_name),
             CheckErrors::UnknownFunction(function_name) => format!("use of unresolved function '{}'", function_name),
             CheckErrors::WriteAttemptedInReadOnly => format!("expecting read-only statements, detected a writing operation"),
-            CheckErrors::BadAssetName => format!("expecting an asset name as an argument"),
             CheckErrors::BadTokenName => format!("expecting an token name as an argument"),
-            CheckErrors::DefineTokenBadSignature => format!("(define-token ...) expects a token name as an argument"),
-            CheckErrors::DefineAssetBadSignature => format!("(define-asset ...) expects an asset name and an asset identifier type signature as arguments"),
-            CheckErrors::NoSuchAsset(asset_name) => format!("tried to use asset function with a undefined asset ('{}')", asset_name),
-            CheckErrors::NoSuchToken(asset_name) => format!("tried to use token function with a undefined token ('{}')", asset_name),
+            CheckErrors::DefineFTBadSignature => format!("(define-token ...) expects a token name as an argument"),
+            CheckErrors::DefineNFTBadSignature => format!("(define-asset ...) expects an asset name and an asset identifier type signature as arguments"),
+            CheckErrors::NoSuchNFT(asset_name) => format!("tried to use asset function with a undefined asset ('{}')", asset_name),
+            CheckErrors::NoSuchFT(asset_name) => format!("tried to use token function with a undefined token ('{}')", asset_name),
             CheckErrors::TypeAlreadyAnnotatedFailure | CheckErrors::CheckerImplementationFailure | CheckErrors::NotImplemented => {
                 format!("internal error - please file an issue on github.com/blockstack/blockstack-core")
             },
