+++ src/vm/checker/typecheck/natives/assets.rs
@@ -7,13 +7,13 @@ pub fn check_special_get_owner(checker: &mut TypeChecker, args: &[SymbolicExpres
     check_argument_count(2, args)?;
 
     let asset_name = args[0].match_atom()
-        .ok_or(CheckErrors::BadAssetName)?;
+        .ok_or(CheckErrors::BadTokenName)?;
     checker.type_map.set_type(&args[0], no_type())?;
 
     let supplied_asset_type = checker.type_check(&args[1], context)?;
 
-    let expected_asset_type = checker.contract_context.get_asset_type(asset_name)
-        .ok_or(CheckErrors::NoSuchAsset(asset_name.clone()))?;
+    let expected_asset_type = checker.contract_context.get_nft_type(asset_name)
+        .ok_or(CheckErrors::NoSuchNFT(asset_name.clone()))?;
 
     if !expected_asset_type.admits_type(&supplied_asset_type) {
         Err(CheckErrors::TypeError(expected_asset_type.clone(), supplied_asset_type).into())
@@ -30,8 +30,8 @@ pub fn check_special_get_balance(checker: &mut TypeChecker, args: &[SymbolicExpr
         .ok_or(CheckErrors::BadTokenName)?;
     checker.type_map.set_type(&args[0], no_type())?;
 
-    if !checker.contract_context.token_exists(asset_name) {
-        return Err(CheckErrors::NoSuchToken(asset_name.clone()).into());
+    if !checker.contract_context.ft_exists(asset_name) {
+        return Err(CheckErrors::NoSuchFT(asset_name.clone()).into());
     }
 
     let expected_owner_type: TypeSignature = AtomTypeIdentifier::PrincipalType.into();
@@ -44,12 +44,12 @@ pub fn check_special_mint_asset(checker: &mut TypeChecker, args: &[SymbolicExpre
     check_argument_count(3, args)?;
 
     let asset_name = args[0].match_atom()
-        .ok_or(CheckErrors::BadAssetName)?;
+        .ok_or(CheckErrors::BadTokenName)?;
     checker.type_map.set_type(&args[0], no_type())?;
 
     let expected_owner_type: TypeSignature = AtomTypeIdentifier::PrincipalType.into();
-    let expected_asset_type = checker.contract_context.get_asset_type(asset_name)
-        .ok_or(CheckErrors::NoSuchAsset(asset_name.clone()))?
+    let expected_asset_type = checker.contract_context.get_nft_type(asset_name)
+        .ok_or(CheckErrors::NoSuchNFT(asset_name.clone()))?
         .clone(); // this clone shouldn't be strictly necessary, but to use `type_check_expects` with this, it would have to be.
 
     checker.type_check_expects(&args[1], context, &expected_asset_type)?;
@@ -74,8 +74,8 @@ pub fn check_special_mint_token(checker: &mut TypeChecker, args: &[SymbolicExpre
     checker.type_check_expects(&args[2], context, &expected_owner_type)?;
 
 
-    if !checker.contract_context.token_exists(asset_name) {
-        return Err(CheckErrors::NoSuchToken(asset_name.clone()).into());
+    if !checker.contract_context.ft_exists(asset_name) {
+        return Err(CheckErrors::NoSuchFT(asset_name.clone()).into());
     }
     
     Ok(AtomTypeIdentifier::ResponseType(
@@ -86,13 +86,13 @@ pub fn check_special_mint_token(checker: &mut TypeChecker, args: &[SymbolicExpre
 pub fn check_special_transfer_asset(checker: &mut TypeChecker, args: &[SymbolicExpression], context: &TypingContext) -> TypeResult {
     check_argument_count(4, args)?;
 
-    let asset_name = args[0].match_atom()
-        .ok_or(CheckErrors::BadAssetName)?;
+    let token_name = args[0].match_atom()
+        .ok_or(CheckErrors::BadTokenName)?;
     checker.type_map.set_type(&args[0], no_type())?;
 
     let expected_owner_type: TypeSignature = AtomTypeIdentifier::PrincipalType.into();
-    let expected_asset_type = checker.contract_context.get_asset_type(asset_name)
-        .ok_or(CheckErrors::NoSuchAsset(asset_name.clone()))?
+    let expected_asset_type = checker.contract_context.get_nft_type(token_name)
+        .ok_or(CheckErrors::NoSuchNFT(token_name.clone()))?
         .clone();
 
     checker.type_check_expects(&args[1], context, &expected_asset_type)?;
@@ -107,7 +107,7 @@ pub fn check_special_transfer_asset(checker: &mut TypeChecker, args: &[SymbolicE
 pub fn check_special_transfer_token(checker: &mut TypeChecker, args: &[SymbolicExpression], context: &TypingContext) -> TypeResult {
     check_argument_count(4, args)?;
 
-    let asset_name = args[0].match_atom()
+    let token_name = args[0].match_atom()
         .ok_or(CheckErrors::BadTokenName)?;
     checker.type_map.set_type(&args[0], no_type())?;
 
@@ -118,8 +118,8 @@ pub fn check_special_transfer_token(checker: &mut TypeChecker, args: &[SymbolicE
     checker.type_check_expects(&args[2], context, &expected_owner_type)?; // owner
     checker.type_check_expects(&args[3], context, &expected_owner_type)?; // recipient
 
-    if !checker.contract_context.token_exists(asset_name) {
-        return Err(CheckErrors::NoSuchToken(asset_name.clone()).into());
+    if !checker.contract_context.ft_exists(token_name) {
+        return Err(CheckErrors::NoSuchFT(token_name.clone()).into());
     }
 
     Ok(AtomTypeIdentifier::ResponseType(
