+++ src/vm/checker/typecheck/tests/assets.rs
@@ -120,39 +120,39 @@ fn test_bad_asset_usage() {
     ];
 
     let expected = [
-        CheckErrors::NoSuchToken("stackoos".to_string()),
+        CheckErrors::NoSuchFT("stackoos".to_string()),
         CheckErrors::BadTokenName,
         CheckErrors::TypeError(AtomTypeIdentifier::PrincipalType.into(),
                                AtomTypeIdentifier::IntType.into()),
-        CheckErrors::BadAssetName,
-        CheckErrors::NoSuchAsset("stackoos".to_string()),
+        CheckErrors::BadTokenName,
+        CheckErrors::NoSuchNFT("stackoos".to_string()),
         CheckErrors::TypeError(AtomTypeIdentifier::BufferType(10).into(),
                                AtomTypeIdentifier::IntType.into()),
         CheckErrors::TypeError(AtomTypeIdentifier::BufferType(10).into(),
                                AtomTypeIdentifier::BufferType(15).into()),
-        CheckErrors::BadAssetName,
-        CheckErrors::NoSuchAsset("stackoos".to_string()),
+        CheckErrors::BadTokenName,
+        CheckErrors::NoSuchNFT("stackoos".to_string()),
         CheckErrors::TypeError(AtomTypeIdentifier::BufferType(10).into(),
                                AtomTypeIdentifier::IntType.into()),
         CheckErrors::TypeError(AtomTypeIdentifier::BufferType(10).into(),
                                AtomTypeIdentifier::BufferType(15).into()),
         CheckErrors::TypeError(AtomTypeIdentifier::PrincipalType.into(),
                                AtomTypeIdentifier::IntType.into()),
-        CheckErrors::NoSuchToken("stackoos".to_string()),
+        CheckErrors::NoSuchFT("stackoos".to_string()),
         CheckErrors::BadTokenName,
         CheckErrors::TypeError(AtomTypeIdentifier::PrincipalType.into(),
                                AtomTypeIdentifier::IntType.into()),
         CheckErrors::TypeError(AtomTypeIdentifier::IntType.into(),
                                AtomTypeIdentifier::BoolType.into()),
-        CheckErrors::BadAssetName,
-        CheckErrors::NoSuchAsset("stackoos".to_string()),
+        CheckErrors::BadTokenName,
+        CheckErrors::NoSuchNFT("stackoos".to_string()),
         CheckErrors::TypeError(AtomTypeIdentifier::PrincipalType.into(),
                                AtomTypeIdentifier::IntType.into()),
         CheckErrors::TypeError(AtomTypeIdentifier::PrincipalType.into(),
                                AtomTypeIdentifier::IntType.into()),
         CheckErrors::TypeError(AtomTypeIdentifier::BufferType(10).into(),
                                AtomTypeIdentifier::IntType.into()),
-        CheckErrors::NoSuchToken("stackoos".to_string()),
+        CheckErrors::NoSuchFT("stackoos".to_string()),
         CheckErrors::BadTokenName,
         CheckErrors::TypeError(AtomTypeIdentifier::PrincipalType.into(),
                                AtomTypeIdentifier::IntType.into()),
@@ -162,7 +162,7 @@ fn test_bad_asset_usage() {
                                AtomTypeIdentifier::IntType.into()),
         CheckErrors::TypeError(AtomTypeIdentifier::IntType.into(),
                                AtomTypeIdentifier::BoolType.into()),
-        CheckErrors::DefineAssetBadSignature.into(),
+        CheckErrors::DefineNFTBadSignature.into(),
     ];
 
     let mut db = AnalysisDatabase::memory();
